Refactor commands provider
